+++ config/default.php
@@ -38,7 +38,7 @@ $config['calibre_internal_directory'] = '';
  * Note: not all features will be available in this configuration, e.g.
  * no kepub version or metadata updates, and no monocle epub reader
  * Example (test):
- * $config['calibre_external_storage'] = 'http://localhost/cops/test/BaseWithSomeBooks/';
+ * $config['calibre_external_storage'] = 'http://localhost/cops/tests/BaseWithSomeBooks/';
  */
 $config['calibre_external_storage'] = '';