+++ CHANGELOG.md
@@ -14,6 +14,9 @@ x.x.x - TODO
   * ...
 
 3.x.x - 2024xxxx
+  * ...
+
+3.1.1 - 20240909 Release candidate for 3.x
   * Changes in config/default.php file:
     - new $config['cops_front_controller'] = '' value
   * Set front controller to remove index.php/ from route URLs